refactor: :recycle: moves utility functions to service layer
